Added new view for cart table

+++ web-api/src/main/java/com/app/model/cart/CartModel.java
@@ -10,11 +10,11 @@ import java.math.BigDecimal;
 public class CartModel {
     @Id @Column(name = "user_id")    private String  userId;
     @Id @Column(name = "product_id") private Integer productId;
-    @Column(name = "quantity") private BigDecimal quantity;
+    @Column(name = "quantity") private Long quantity;
 
     //Constructors
     public CartModel(){}
-    public CartModel(String userId, Integer productId, BigDecimal quantity) {
+    public CartModel(String userId, Integer productId, Long quantity) {
         this.userId = userId;
         this.productId = productId;
         this.quantity = quantity;
@@ -28,8 +28,8 @@ public class CartModel {
     public Integer getProductId() { return productId; }
     public void setProductId(Integer productId) { this.productId = productId; }
 
-    public BigDecimal getQuantity() { return quantity; }
-    public void setQuantity(BigDecimal quantity) { this.quantity = quantity; }
+    public Long getQuantity() { return quantity; }
+    public void setQuantity(Long quantity) { this.quantity = quantity; }
 
     static class CartPrimaryKeys implements Serializable {
         private String   userId;
